+++ src/react-router/GoogleAnalyticsProvider.tsx
