+++ src/react-router/index.ts
@@ -1,0 +1,2 @@
+export {default as GoogleAnalyticsProvider} from './GoogleAnalyticsProvider';
+export {default as LinkBehavior} from './LinkBehavior';
